validation
added validation for all commands&queries with FluentValidation

+++ Petsitters.Backend/Petsitters.Application/Feedbacks/Commands/CreateFeedback/CreateFeedbackCommand.cs
@@ -5,7 +5,7 @@ namespace Petsitters.Application.Feedbacks.Commands.CreateFeedback
     public class CreateFeedbackCommand  : IRequest<int>
     {
         public string Name { get; set; }
-        public short Mark { get; set; }
+        public int Mark { get; set; }
         public string Text { get; set; }
         public int OwnerUserId { get; set; }
         public int WorkerUserId { get; set; }
