+++ Petsitters.Backend/Petsitters.Domain/Feedback.cs
@@ -5,7 +5,7 @@ namespace Petsitters.Domain
     public class Feedback
     {
         public int Id { get; set; }
-        public short Mark { get; set; }
+        public int Mark { get; set; }
         public string Text { get; set; }
         public int OwnerUserId { get; set; }
         public User Owner { get; set; }
